Kotlin 1.5 (JS failing on :kodein-memory-crypto:jsBrowserTest)

+++ kodein-memory-crypto/src/jvmMain/kotlin/org/kodein/memory/crypto/Pbkdf2Jvm.kt
@@ -21,7 +21,7 @@ public actual object Pbkdf2 {
         val derivedKey = SecretKeyFactory.getInstance(algorithmName)
             .generateSecret(
                 PBEKeySpec(
-                    CharArray(password.size) { password[it].toChar() },
+                    CharArray(password.size) { password[it].toInt().toChar() },
                     salt.getBytes(),
                     rounds,
                     dst.size * 8
